Switch org to com.github.sbt
