+++ build.sbt
@@ -1,7 +1,7 @@
 name := "sbt-findbugs"
-organization := "org.scala-sbt"
+organization := "com.github.sbt"
 
-version := "2.0.0-M3"
+version := "2.0.0-M4-SNAPSHOT"
 
 sbtPlugin := true
 crossSbtVersions := Seq("0.13.16", "1.0.2")
