+++ src/main/scala/com/github/sbt/findbugs/FindbugsRunner.scala
@@ -10,7 +10,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.scalasbt.findbugs
+package com.github.sbt.findbugs
 
 import java.io.File
