+++ src/sbt-test/sbt-findbugs/settings/build.sbt
