+++ src/sbt-test/sbt-findbugs/settings/src/main/java/com/example/Bad.java
