+++ src/sbt-test/sbt-findbugs/settings/test
