+++ src/sbt-test/sbt-findbugs/simple/build.sbt
