+++ src/sbt-test/sbt-findbugs/simple/src/main/java/com/example/Bad.java
